% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_net.R
\name{deal_net}
\alias{deal_net}
\title{Title deal with network}
\usage{
deal_net(net, dise_gene)
}
\arguments{
\item{net}{a network}

\item{dise_gene}{a matrix with one column of genes}
}
\value{
a matrix
}
\description{
Title deal with network
}
\examples{
deal_net(net,dise_gene)
}
