% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edge_weight}
\alias{edge_weight}
\title{weights of edges of a net}
\format{A matrix with 25 rows and 3 columns}
\usage{
edge_weight
}
\description{
the first two columns are a net, and third column is their weight
}
\keyword{datasets}
