% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_mat}
\alias{genes_mat}
\title{a one-to-many matrix of GO term and gene}
\format{A matrix with 45 rows and 3 columns}
\usage{
genes_mat
}
\description{
the first column is the gene symbol, the second column is the go terms
}
\details{
the third column is the number of go terms
}
\keyword{datasets}
