% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{get_edge_weight}
\alias{get_edge_weight}
\title{Title weight edge}
\usage{
get_edge_weight(net_disease_term)
}
\arguments{
\item{net_disease_term}{GO terms for each pair of nodes in the network}
}
\value{
a matrix
}
\description{
Title weight edge
}
\examples{
get_edge_weight(net_disease_term)
}
