% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{get_neighbor}
\alias{get_neighbor}
\title{Title get neighbor of a node}
\usage{
get_neighbor(node, net)
}
\arguments{
\item{node}{a gene}

\item{net}{a network}
}
\value{
a vector of gene
}
\description{
Title get neighbor of a node
}
