% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_GO.R
\name{get_net_disease_term}
\alias{get_net_disease_term}
\title{Title Get the GO terms for each pair of nodes in the network}
\usage{
get_net_disease_term(genes_mat, net_disease)
}
\arguments{
\item{genes_mat}{a one-to-many matrix of GO term and gene}

\item{net_disease}{a disease related network, matrix}
}
\value{
a matrix
}
\description{
Title Get the GO terms for each pair of nodes in the network
}
\examples{
get_net_disease_term(genes_mat,net_disease)
}
