% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{get_node_weight}
\alias{get_node_weight}
\title{Title weight node}
\usage{
get_node_weight(genes_mat)
}
\arguments{
\item{genes_mat}{a one-to-many matrix of GO term and gene}
}
\value{
a matrix
}
\description{
Title weight node
}
\examples{
get_node_weight(genes_mat)
}
