% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{get_Q}
\alias{get_Q}
\title{Title get the disease risk transition probability matrix}
\usage{
get_Q()
}
\value{
a matrix
}
\description{
Title get the disease risk transition probability matrix
}
\examples{
get_Q()
}
