% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{get_R}
\alias{get_R}
\title{Title get the final genetic disease risk scores}
\usage{
get_R(Q, bet, R_0, node_weight, threshold = 10^(-9))
}
\arguments{
\item{Q}{the disease risk transition probability matrix}

\item{bet}{a parameter to measure the importance of genes and interactions}

\item{R_0}{the vector of initial disease risk scores for all genes}

\item{node_weight}{a matrix, genes and their weights}

\item{threshold}{a threshold for terminating iterations}
}
\value{
a matrix
}
\description{
Title get the final genetic disease risk scores
}
\examples{
get_R(Q,0.5,R_0,node_weight,10^(-9))
}
