% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{get_R_0}
\alias{get_R_0}
\title{Title get the vector of initial disease risk scores for all genes}
\usage{
get_R_0(disease_gene, node_weight, f = 1)
}
\arguments{
\item{disease_gene}{a matrix of a column of genes}

\item{node_weight}{a matrix, genes and their weights}

\item{f}{an integer parameter to measure the significance of disease
genes and candidate genes}
}
\value{
a vector
}
\description{
Title get the vector of initial disease risk scores for all genes
}
\examples{
get_R_0(dise_gene,node_weight,1)
}
