% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_GO.R
\name{get_term_mat}
\alias{get_term_mat}
\title{Title Get a one-to-many matrix of GO term and gene}
\usage{
get_term_mat(net_disease, GO_human)
}
\arguments{
\item{net_disease}{a disease related network, matrix}

\item{GO_human}{a matrix, gene and GO terms}
}
\value{
a matrix
}
\description{
Title Get a one-to-many matrix of GO term and gene
}
\examples{
get_term_mat(net_disease,GO_human)
}
