% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_human}
\alias{GO_human}
\title{a matrix, gene and GO terms}
\format{A matrix with 481418 rows and 2 column}
\usage{
GO_human
}
\description{
the first column is the gene symbol, the second column is the go terms
}
\keyword{datasets}
