% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolic_net}
\alias{metabolic_net}
\title{a matrix, Human metabolic network}
\format{A matrix with 589,199 rows and 2 columns}
\usage{
metabolic_net
}
\description{
a matrix, Human metabolic network
}
\keyword{datasets}
