% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_disease}
\alias{net_disease}
\title{a network of disease related genes}
\format{A matrix with 26 rows and 2 columns}
\usage{
net_disease
}
\description{
a network of disease related genes
}
\keyword{datasets}
