% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_disease_term}
\alias{net_disease_term}
\title{GO terms for each pair of nodes in the network}
\format{A matrix with 25 rows and 4 columns}
\usage{
net_disease_term
}
\description{
the first two columns is the network
}
\details{
the third column is the go terms,the fourth column is the number of go terms

the fourth column is the number of go terms
}
\keyword{datasets}
