% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{node_weight}
\alias{node_weight}
\title{a matrix, genes and their weights}
\format{A matrix with 45 rows and 2 columns}
\usage{
node_weight
}
\description{
a matrix, genes and their weights
}
\keyword{datasets}
