% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Q}
\alias{Q}
\title{the disease risk transition probability matrix}
\format{A matrix with 45 rows and 45 column}
\usage{
Q
}
\description{
the disease risk transition probability matrix
}
\keyword{datasets}
