% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_0}
\alias{R_0}
\title{the vector of initial disease risk scores for all genes}
\format{A vector of 45 number}
\usage{
R_0
}
\description{
the vector of initial disease risk scores for all genes
}
\keyword{datasets}
