% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terms_mat}
\alias{terms_mat}
\title{a matrix, GO terms and GO genes}
\format{A matrix with 1172 rows and 3 columns}
\usage{
terms_mat
}
\description{
a one-to-many matrix of GO term and gene
}
\keyword{datasets}
