% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_GO.R
\name{get_term_mat}
\alias{get_term_mat}
\title{Get a one-to-many matrix of GO term and gene}
\usage{
get_term_mat(net_disease)
}
\arguments{
\item{net_disease}{a disease related network, matrix}
}
\value{
a matrix
}
\description{
Get a one-to-many matrix of GO term and gene
}
\examples{
get_term_mat(net_disease)
}
